[005] add new file: home.dart, separate codes ...
